Add ComponentKey
